fix scroll by line number

+++ src/app.rs
@@ -15,7 +15,7 @@ pub struct App<'a> {
     pub records_buf: Vec<fastq::Record>,
     pub line_buf: Vec<Line<'a>>,
     pub mode: UIMode,
-    pub line_num: u16,
+    pub line_num: usize,
     file: String,
     reader: Reader<std::io::BufReader<std::fs::File>>, // buf_size
 }
